questionAI/requirements.txt
streamlit
streamlit-authenticator
PyYAML
bcrypt
pillow
google-generativeai
google-genai
numpy
fastapi
uvicorn

